make FakeAlertToast always been build
