+++ Sources/SwiftyAlert/AlertToast+/FakeAlertToast.swift
@@ -1,0 +1,107 @@
+//
+//  FakeAlertToast.swift
+//  
+//
+//  Created by Dove Zachary on 2024/2/20.
+//
+
+import SwiftUI
+
+// This is because Xcode can not recognize #if canImport(AlertToast) when add package from url.
+public struct FakeAlertToast {
+    /// Determine how the alert will be display
+    public enum DisplayMode: Equatable{
+        
+        ///Present at the center of the screen
+        case alert
+        
+        ///Drop from the top of the screen
+        case hud
+        
+        ///Banner from the bottom of the view
+        case banner(_ transition: BannerAnimation)
+    }
+    
+    /// Determine what the alert will display
+    public enum AlertType: Equatable{
+        
+        ///Animated checkmark
+        case complete(_ color: Color)
+        
+        ///Animated xmark
+        case error(_ color: Color)
+        
+        ///System image from `SFSymbols`
+        case systemImage(_ name: String, _ color: Color)
+        
+        ///Image from Assets
+        case image(_ name: String, _ color: Color)
+        
+        ///Loading indicator (Circular)
+        case loading
+        
+        ///Only text alert
+        case regular
+    }
+    
+    public enum BannerAnimation{
+        case slide, pop
+    }
+    public enum AlertStyle: Equatable{
+        
+        case style(backgroundColor: Color? = nil,
+                   titleColor: Color? = nil,
+                   subTitleColor: Color? = nil,
+                   titleFont: Font? = nil,
+                   subTitleFont: Font? = nil)
+        
+        ///Get background color
+        var backgroundColor: Color? {
+            switch self{
+            case .style(backgroundColor: let color, _, _, _, _):
+                return color
+            }
+        }
+        
+        /// Get title color
+        var titleColor: Color? {
+            switch self{
+            case .style(_,let color, _,_,_):
+                return color
+            }
+        }
+        
+        /// Get subTitle color
+        var subtitleColor: Color? {
+            switch self{
+            case .style(_,_, let color, _,_):
+                return color
+            }
+        }
+        
+        /// Get title font
+        var titleFont: Font? {
+            switch self {
+            case .style(_, _, _, titleFont: let font, _):
+                return font
+            }
+        }
+        
+        /// Get subTitle font
+        var subTitleFont: Font? {
+            switch self {
+            case .style(_, _, _, _, subTitleFont: let font):
+                return font
+            }
+        }
+    }
+    
+    public init() {}
+    public init(
+        displayMode: DisplayMode = .alert,
+        type: AlertType,
+        title: String? = nil,
+        subTitle: String? = nil,
+        style: AlertStyle? = nil
+    ) {}
+}
